chore: Refactoring progress
