+++ src/Uno.UI/Microsoft/UI/Xaml/Controls/IconSource/ImageIconSource.cs
@@ -6,41 +6,40 @@ using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml.Media;
 
-namespace Microsoft.UI.Xaml.Controls
+namespace Microsoft.UI.Xaml.Controls;
+
+public partial class ImageIconSource : IconSource
 {
-	public partial class ImageIconSource : IconSource
+	public ImageSource ImageSource
 	{
-		public ImageSource ImageSource
+		get => (ImageSource)GetValue(ImageSourceProperty);
+		set => SetValue(ImageSourceProperty, value);
+	}
+
+	public static DependencyProperty ImageSourceProperty { get; } =
+		DependencyProperty.Register(nameof(ImageSource), typeof(ImageSource), typeof(ImageIconSource), new FrameworkPropertyMetadata(null, OnPropertyChanged));
+
+	private protected override IconElement CreateIconElementCore()
+	{
+		var imageIcon = new ImageIcon();
+		if (ImageSource is { } imageSource)
 		{
-			get => (ImageSource)GetValue(ImageSourceProperty);
-			set => SetValue(ImageSourceProperty, value);
+			imageIcon.Source = imageSource;
+		}
+		if (Foreground is { } newForeground)
+		{
+			imageIcon.Foreground = newForeground;
+		}
+		return imageIcon;
+	}
+
+	private protected override DependencyProperty GetIconElementPropertyCore(DependencyProperty sourceProperty)
+	{
+		if (sourceProperty == ImageSourceProperty)
+		{
+			return ImageIcon.SourceProperty;
 		}
 
-		public static DependencyProperty ImageSourceProperty { get; } =
-			DependencyProperty.Register(nameof(ImageSource), typeof(ImageSource), typeof(ImageIconSource), new FrameworkPropertyMetadata(null, OnPropertyChanged));
-
-		private protected override IconElement CreateIconElementCore()
-		{
-			var imageIcon = new ImageIcon();
-			if (ImageSource is { } imageSource)
-			{
-				imageIcon.Source = imageSource;
-			}
-			if (Foreground is { } newForeground)
-			{
-				imageIcon.Foreground = newForeground;
-			}
-			return imageIcon;
-		}
-
-		private protected override DependencyProperty GetIconElementPropertyCore(DependencyProperty sourceProperty)
-		{
-			if (sourceProperty == ImageSourceProperty)
-			{
-				return ImageIcon.SourceProperty;
-			}
-
-			return base.GetIconElementPropertyCore(sourceProperty);
-		}
+		return base.GetIconElementPropertyCore(sourceProperty);
 	}
 }
