+++ src/Uno.UI/UI/Xaml/Controls/Icons/IconSourceElement.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Uno.UI.UI.Xaml.Controls.Icons
+{
+	internal class IconSourceElement
+	{
+	}
+}
